+++ src/pizza/dtos/baked-repo.dto.ts
@@ -12,7 +12,7 @@ export class BakeRepoDto {
 
   @ApiProperty({
     description: "Option to wait for Pizza service to finish baking repo",
-    type: String,
+    type: Boolean,
     example: false,
   })
   @IsBoolean()